% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add}
\alias{add}
\title{Adds small spatiotemporal map (stm) to the big stm at specified point}
\usage{
add(big, a, p, checks = T)
}
\arguments{
\item{big}{spatiotemporal map}

\item{a}{small spatiotemporal map which will be added}

\item{p}{coordinates where wil be stm a added to big}

\item{checks}{boolean, if we should run tests. We might want to set this to F to reduce running time. Default value is T}
}
\description{
Adds small spatiotemporal map (stm) to the big stm at specified point
}
\author{
Filip Dechterenko
}

