% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.scanpath}
\alias{as.scanpath}
\title{Converts an eye to the scanpath}
\usage{
as.scanpath(eye)
}
\arguments{
\item{eye}{an object of class eye}
}
\description{
Converts an object of class eye to the object of class scanpath
}
\author{
Filip Dechterenko
}

