% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.valid.eyes.data.frame}
\alias{as.valid.eyes.data.frame}
\title{Converts data frame to valid.eyes data.frame}
\usage{
as.valid.eyes.data.frame(df)
}
\arguments{
\item{df}{data.frame which should contain valid eyes}
}
\description{
Converts data frame to valid.eyes data.frame. Slices data frame per each pair id, trial and tries to convert each slice into eye object. If all conversions succeed, class valid.eyes is added to the data frame. If data.frame could not be converted, method returns NULL
}
\author{
Filip Dechterenko
}

