% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bin.trajectory}
\alias{bin.trajectory}
\title{Bins trajectory}
\usage{
bin.trajectory(spath)
}
\arguments{
\item{spath}{scanpath object which will be binned into discrete spatiotemporal map.}
}
\description{
Bins trajectory to bins specified by parameters in config
}
\author{
Filip Dechterenko
}

