% Generated by roxygen2 (4.0.1): do not edit by hand
\name{center.axis}
\alias{center.axis}
\title{Centers two spatiotemporal map at each other}
\usage{
center.axis(a1, a2, p1, p2)
}
\arguments{
\item{a1}{first spatiotemporal map}

\item{a2}{second spatiotemporal map}

\item{p1}{point in first spatiotemporal map}

\item{p2}{point in second spatiotemporal map}
}
\description{
Centers two spatiotemporal map at each other.Returns part of a2, when a1 is placed over it, p1 over p2
}
\author{
Filip Dechterenko
}

