% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clear.space}
\alias{clear.space}
\title{Removes all data from spatiotemporal map}
\usage{
clear.space(space)
}
\arguments{
\item{space}{spatiotemporal map}
}
\description{
Removes all data from spatiotemporal map
}
\author{
Filip Dechterenko
}

