% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compute.nss.from.space}
\alias{compute.nss.from.space}
\title{Computes nss distance metric}
\usage{
compute.nss.from.space(V, SV)
}
\arguments{
\item{V}{scanpathwhich will be tested}

\item{SV}{stm map}
}
\description{
Computes nss distance metric for eye object against spatiotemporal map
}
\author{
Filip Dechterenko
}

