% Generated by roxygen2 (4.0.1): do not edit by hand
\name{contains}
\alias{contains}
\title{Tests, if coordinates of the bin are present in spatiotemporal map}
\usage{
contains(space, point)
}
\arguments{
\item{space}{spatiotemporal map}

\item{point}{vector of coordinates of the bin}
}
\description{
Tests, if coordinates of the bin are present in spatiotemporal map
}
\author{
Filip Dechterenko
}

