% Generated by roxygen2 (4.0.1): do not edit by hand
\name{corr.dist}
\alias{corr.dist}
\title{Computes correlation distance metric}
\usage{
corr.dist(SV1,SV2)
corr.dist(eye1,eye2)
}
\arguments{
\item{SV1}{first spatiotemporal map}

\item{SV2}{second spatiotemporal map}

\item{eye1}{first eye gaze data}

\item{eye2}{second eye gaze data}
}
\description{
Computes correlation distance metric between two spatiotemporal maps or two eye gaze data (in second case corresponding stm are loaded)
}
\author{
Filip Dechterenko
}

