% Generated by roxygen2 (4.0.1): do not edit by hand
\name{corr.for.two.list}
\alias{corr.for.two.list}
\title{Computes correlation distance metric for two lists}
\usage{
corr.for.two.list(list1, list2, spaces.dir = NULL)
}
\arguments{
\item{list1}{first list with stm or ids in format id.trial}

\item{list2}{second list with stm or ids in format id.trial}

\item{spaces.dir}{optional parameter with path to the stored stm maps. If missing, default value is used from config}
}
\description{
Computes correlation distance metric between two lists of spatiotemporal maps (stm) or ids of eye gaze data.
In second case corresponding stm are loaded from path specified in config or in parameter
}
\author{
Filip Dechterenko
}

