% Generated by roxygen2 (4.0.1): do not edit by hand
\name{describe.eye}
\alias{describe.eye}
\title{Descriptive statistics for an eye object}
\usage{
describe.eye(eye)
}
\arguments{
\item{eye}{an object of class eye}
}
\description{
Descriptive statistics for an eye object. It produces min, max in x and y coordinates. Mean differences between time points and convex hull
}
\author{
Filip Dechterenko
}

