% Generated by roxygen2 (4.0.1): do not edit by hand
\name{empty.space}
\alias{empty.space}
\title{Creates empty spatiotemporal map}
\usage{
empty.space(xsteps = NULL, ysteps = NULL, tsteps = NULL)
}
\arguments{
\item{xsteps}{steps for spatiotemporal map in x coordinate}

\item{ysteps}{steps for spatiotemporal map in y coordinate}

\item{tsteps}{steps for spatiotemporal map in t coordinate}
}
\description{
Creates empty spatiotemporal map with specified steps
}
\author{
Filip Dechterenko
}

