% Generated by roxygen2 (4.0.1): do not edit by hand
\name{flip.eye}
\alias{flip.eye}
\title{Flips eye by x and/or y axis}
\usage{
flip.eye(eye, flip.x = F, flip.y = F)
}
\arguments{
\item{eye}{an object of class eye}

\item{flip.x}{logical, if we should flip x coord}

\item{flip.y}{logical, if we should flip y coord}
}
\description{
Flips eye by x and/or y axis
}
\author{
Filip Dechterenko
}

