% Generated by roxygen2 (4.0.1): do not edit by hand
\name{flip.scanpath}
\alias{flip.scanpath}
\title{Flips scanpath by x and/or y axis}
\usage{
flip.scanpath(spath, flip.x = F, flip.y = F)
}
\arguments{
\item{spath}{scanpath object which will be flipped}

\item{flip.x}{logical, if we should flip x axis}

\item{flip.y}{logical, if we should flip y axis}
}
\description{
Flips scanpath by x and/or y axis
}
\author{
Filip Dechterenko
}

