% Generated by roxygen2 (4.0.1): do not edit by hand
\name{flip.space}
\alias{flip.space}
\title{Flips spatiotemporal map by x and/or y axis}
\usage{
flip.space(V, type = 1)
}
\arguments{
\item{V}{spatiotemporal object which will be flipped}

\item{type}{variable determining, around which axis we will flip map.}
}
\description{
Flips spatiotemporal map by x or y axis
}
\author{
Filip Dechterenko
}

