% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gaussian.mask}
\alias{gaussian.mask}
\title{Creates small gaussian}
\usage{
gaussian.mask(steps = NULL, xsd = NULL, ysd = NULL, tsd = NULL)
}
\arguments{
\item{steps}{objects with step sizes for individual dimension}

\item{xsd}{standard deviation of gaussian in x coord. Default value is taken from config}

\item{ysd}{standard deviation of gaussian in y coord. Default value is taken from config}

\item{tsd}{standard deviation of gaussian in t coord. Default value is taken from config}
}
\description{
Creates small gaussian in form of spatiotemporal fixation map
}
\author{
Filip Dechterenko
}

