% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.coords}
\alias{get.coords}
\title{Tests if object is of class space}
\usage{
get.coords(space, index, values = T)
}
\arguments{
\item{obj}{an object to be tested}
}
\description{
Simply test, if object is of class space
}
\author{
Filip Dechterenko
}

