% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.eye}
\alias{get.eye}
\title{Creates eye object from data frame}
\usage{
get.eye(id, trial, eye.data)
}
\arguments{
\item{id}{integer of participant id}

\item{trial}{integer of trial number}

\item{eye.data}{data.frame of class valid.eyes}
}
\description{
We need to have data.frame with eye data. This function selects data from data.frame specified by trial and id
}
\author{
Filip Dechterenko
}

