% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.scanpath}
\alias{get.scanpath}
\title{Gets scanpaths from data.frame}
\usage{
get.scanpath(id, trial, eyeData)
}
\arguments{
\item{id}{integer with participant id}

\item{trial}{integer with trial id}

\item{eyeData}{data.frame with eye gaze data}
}
\description{
Selects scanpath from data.frame based on id and trial
}
\author{
Filip Dechterenko
}

