% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.stored.scanpath.name}
\alias{get.stored.scanpath.name}
\title{Gets filename of stored scanpath}
\usage{
get.stored.scanpath.name(id, trial, prefix = "", spaces.dir = NULL)
}
\arguments{
\item{id}{integer with participant id}

\item{trial}{integer with trial id}

\item{prefix}{optional prefix of the scanpath}

\item{space.dir}{directory with stored scanpaths. Default value is taken from config}
}
\description{
Construct filename for stored scanpath based on id, trial, prefix and directory, where are scanpaths stored
}
\author{
Filip Dechterenko
}

