% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.stored.space.name}
\alias{get.stored.space.name}
\title{Gets filename of stored spatiotemporal map}
\usage{
get.stored.space.name(id, trial, prefix, spaces.dir = NULL)
}
\arguments{
\item{id}{integer with participant id}

\item{trial}{integer with trial id}

\item{prefix}{optional prefix of the spatiotemporal map}

\item{space.dir}{directory with stored spatiotemporal maps. Default value is taken from config}
}
\description{
Construct filename for stored spatiotemporal map based on id, trial, prefix and directory, where are spatiotemporal maps stored
}
\author{
Filip Dechterenko
}

