% Generated by roxygen2 (4.0.1): do not edit by hand
\name{incoherence.eye}
\alias{incoherence.eye}
\title{Randomly move proportion of trajectory far away}
\usage{
incoherence.eye(eye, ratio)
}
\arguments{
\item{eye}{an object of class eye}

\item{ratio}{fraction of trajectory which should be moved away}
}
\description{
Randomly move proportion of trajectory far away
}
\author{
Filip Dechterenko
}

