% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.compatible}
\alias{is.compatible}
\title{Test, if two spatiotemporal map are compatible}
\usage{
is.compatible(a, b)
}
\arguments{
\item{a}{first stm}

\item{a}{second stm}
}
\description{
Test, if two spatiotemporal map (stm) are compatible (if their step sizes are gcd)
}
\author{
Filip Dechterenko
}

