% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.spath}
\alias{is.spath}
\title{Test if object is of class scanpath}
\usage{
is.spath(obj)
}
\arguments{
\item{obj}{object which will be tested for scanpath membership}
}
\description{
Test if object is of class scanpath
}
\author{
Filip Dechterenko
}

