% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.summable}
\alias{is.summable}
\title{Test, if two spatiotemporal map are summable}
\usage{
is.summable(a, b)
}
\arguments{
\item{a}{first stm}

\item{a}{second stm}
}
\description{
Test, if two spatiotemporal map (stm) are summable
}
\author{
Filip Dechterenko
}

