% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.valid.eye}
\alias{is.valid.eye}
\title{Tests if object is a valid instance of eye object}
\usage{
is.valid.eye(eye)
}
\arguments{
\item{eye}{an eye object to be tested}
}
\description{
This function tests, if object eye has a correct structure for an eye object.
It checks following:
* eye gaze data are inside arena
* time vector has same length as eye gaze data
* differences between time samples are all the same
}
\author{
Filip Dechterenko
}

