% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.valid.eyes}
\alias{is.valid.eyes}
\title{Test if object is of class valid.eyes}
\usage{
is.valid.eyes(obj)
}
\arguments{
\item{obj}{object, which should be tested for membership in both classes data.frame and valid.eyes}
}
\description{
Test if object is of class valid.eyes
}
\author{
Filip Dechterenko
}

