% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load.eye}
\alias{load.eye}
\title{Loads an eye object from file}
\usage{
load.eye(id, tr, eye.dir = NULL)
}
\arguments{
\item{id}{integer specifying participant}

\item{tr}{integer specifying trial}

\item{eye.dir}{optional parametr specifying path to the eye data. If missing, default value is used}
}
\description{
Loads eye object from RData file specified by id and trial. Pth to the directory with eye objects can be specified, or it uses default value from global parameters
}
\author{
Filip Dechterenko
}

