% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load.eye2}
\alias{load.eye2}
\title{Loads an eye object from file}
\usage{
load.eye2(filename)
}
\arguments{
\item{filename}{path to the eye data}
}
\description{
Loads eye object from RData file specified by the path.
}
\author{
Filip Dechterenko
}

