% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load.scanpath}
\alias{load.scanpath}
\title{Loads scanpath from file}
\usage{
load.scanpath(id, trial, spaces.dir)
}
\arguments{
\item{id}{integer of participant id}

\item{trial}{integer of trial id}

\item{spaces.dir}{path to the directory, where scanpaths will be stored. Default value is taken from config}
}
\description{
Loads scanpath from file based on id, trial and optionaly directory, where are scanpaths stored
}
\author{
Filip Dechterenko
}

