% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load.scanpath2}
\alias{load.scanpath2}
\title{Loads scanpath from file}
\usage{
load.scanpath2(filename)
}
\arguments{
\item{filename}{path to the scanpath}
}
\description{
Loads scanpath object in form of discrete spatiotemporal map from RData file specified by the path.
}
\author{
Filip Dechterenko
}

