% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load.space}
\alias{load.space}
\title{Loads spatiotemporal map}
\usage{
load.space(id, trial, prefix = "", spaces.dir = NULL)
}
\arguments{
\item{id}{integer of participant id}

\item{trial}{integer of trial id}

\item{spaces.dir}{path to the directory, where stnm will be stored. Default value is taken from config}
}
\description{
Loads spatiotemporal map (stm) from file based on id, trial and optionaly directory, where are stm stored
}
\author{
Filip Dechterenko
}

