% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load.space2}
\alias{load.space2}
\title{Loads spatiotemporal map from file}
\usage{
load.space2(filename)
}
\arguments{
\item{filename}{path to the spatiotemporal map}
}
\description{
Loads spatiotemporal map object in form of discrete spatiotemporal map from RData file specified by the path.
}
\author{
Filip Dechterenko
}

