% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load.spaces}
\alias{load.spaces}
\title{Loads spaces specified by list}
\usage{
load.spaces(list1)
}
\arguments{
\item{list1}{of ids in format id.trial}
}
\description{
Loads spaces specified by list of values in format id.trial
}
\author{
Filip Dechterenko
}

