% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make.steps.gauss}
\alias{make.steps.gauss}
\title{Creates steps for 3D gaussian}
\usage{
make.steps.gauss(xystep = NULL, tstep = NULL, sdx = NULL, sdy = NULL,
  sdt = NULL)
}
\arguments{
\item{xystep}{size of step in spatial coordinates. Default value is taken from config}

\item{tstep}{size of step in spatial coordinates. Default value is taken from config}

\item{sdx}{standard deviation of gaussian in x coord. Default value is taken from config}

\item{sdy}{standard deviation of gaussian in y coord. Default value is taken from config}

\item{sdt}{standard deviation of gaussian in t coord. Default value is taken from config}
}
\description{
creates steps which will be used for creating 3D gaussian. Those steps will be used for creating spatiotemporal map consisting of gaussian filter
}
\author{
Filip Dechterenko
}

