% Generated by roxygen2 (4.0.1): do not edit by hand
\name{move.eye.polar}
\alias{move.eye.polar}
\title{Moves eye gaze data in space (polar coord)}
\usage{
move.eye.polar(eye, dist = 0, angle = 0, deg = T)
}
\arguments{
\item{eye}{an object of class eye}

\item{dist}{distance from the origin, default=0}

\item{angle}{argument specifying rotation from the top,default=0}

\item{deg}{are we using degrees or radians? Defualt are degrees}
}
\description{
Moves eye gaze data in space. Move is specified by polar coordinates
}
\author{
Filip Dechterenko
}

