% Generated by roxygen2 (4.0.1): do not edit by hand
\name{move.eye}
\alias{move.eye}
\title{Moves eye gaze data in space (cart coord). Move is specified by carthesian coordinates}
\usage{
move.eye(eye, delta.x = 0, delta.y = 0)
}
\arguments{
\item{eye}{an object of class eye}

\item{delta.x}{distance in x coordinate specifying the move in x coord}

\item{delta.y}{distance in y coordinate specifying the move in y coord}
}
\description{
Moves eye gaze data in space (in x and/or y coordinate)
}
\author{
Filip Dechterenko
}

