% Generated by roxygen2 (4.0.1): do not edit by hand
\name{move.space}
\alias{move.space}
\title{Moves spatiotemporal map}
\usage{
move.space(SV, dir.x = 0, dir.y = 0)
}
\arguments{
\item{SV}{spatiotemporal object which will be moevd}

\item{dir.x}{integer meaning how many bins we should move in x axis}

\item{dir.y}{integer meaning how many bins we should move in y axis}
}
\description{
Moves spatiotemporal map by x and/or y coordinates
}
\author{
Filip Dechterenko
}

