% Generated by roxygen2 (4.0.1): do not edit by hand
\name{normalize.space}
\alias{normalize.space}
\title{Normalize values in spatiotemporal map}
\usage{
normalize.space(V)
}
\arguments{
\item{V}{spatiotemporal map}
}
\description{
Normalize values in spatiotemporal map (z-transformation)
}
\author{
Filip Dechterenko
}

