% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nss.dist}
\alias{nss.dist}
\title{Computes nss distance metric}
\usage{
nss.dist(eye1, eye2)
}
\arguments{
\item{eye1}{corresponding scanpath will be loaded for this eye object}

\item{eye2}{corresponding stm will be loaded for this eye object and scanpath will be compared against it}
}
\description{
Computes nss distance metric for one eye object against another
}
\author{
Filip Dechterenko
}

