% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.eye}
\alias{plot.eye}
\title{Visualize an eye object}
\usage{
plot.eye(eye)
}
\arguments{
\item{eye}{an object of class eye}
}
\description{
Visualizes eye gaze data using nice plots from ggplot. Three projections to 2D plane are shown (xy, yt and xt)
}
\author{
Filip Dechterenko
}

