% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.space}
\alias{plot.space}
\title{Visualizes scanpath space}
\usage{
plot.space(space)
}
\arguments{
\item{space}{spatiotemporal map}
}
\description{
Visualizes scanpath space using ggplot. Three projections to 2D plane are shown (xy,yt and xt)
}
\author{
Filip Dechterenko
}

