% Generated by roxygen2 (4.0.1): do not edit by hand
\name{+.space}
\alias{+.space}
\title{Plus operator for two spatiotemporal fixation maps}
\usage{
\method{+}{space}(V1, V2)
}
\arguments{
\item{V1}{first spatiotemporal fixation map}

\item{V2}{second spatiotemporal fixation map}
}
\description{
Plus operator for two spatiotemporal fixation maps
}
\author{
Filip Dechterenko
}

