% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prepare.all.scanpaths}
\alias{prepare.all.scanpaths}
\title{Prepare all scanpaths}
\usage{
prepare.all.scanpaths(response, eyeData, prefix = "", spaces.dir = NULL)
}
\arguments{
\item{response}{data.frame with participant responses}

\item{eyeData}{data.frame with eye gaze data}

\item{prefix}{we can store data with optional prefix, default is ""}

\item{spaces.dir}{path to the directory, where scanpaths will be stored. Default value is taken from config}
}
\description{
Computes all scanpaths and stores them into RData files. Trajectories from response data.frame will be used and data taken from eyeData data.frame
}
\author{
Filip Dechterenko
}

