% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.eye}
\alias{print.eye}
\title{Print an eye object}
\usage{
\method{print}{eye}(eye)
}
\arguments{
\item{eye}{an object of class eye}
}
\description{
Pretty prints eye object, it skips most of the eye gaze data to make it more readable
}
\author{
Filip Dechterenko
}

