% Generated by roxygen2 (4.0.1): do not edit by hand
\name{save.all.eye}
\alias{save.all.eye}
\title{Saves all eye objects from into the file}
\usage{
save.all.eye(df, eye.dir = NULL, verbose = T)
}
\arguments{
\item{df}{data.frame of class valid.eyes}

\item{eye.dir}{optional parametr specifying path to the eye data. If missing, default value from global variables is used.}

\item{verbose}{boolean parameter specifying, if the output should be verbose. Default value is True}
}
\description{
Saves all eye objects from valid.eyes data frame as RData files into directory with eye data
}
\author{
Filip Dechterenko
}

