% Generated by roxygen2 (4.0.1): do not edit by hand
\name{save.eye}
\alias{save.eye}
\title{Saves an eye object into the file}
\usage{
save.eye(eye, eye.dir = NULL)
}
\arguments{
\item{eye}{object to be stored}

\item{eye.dir}{optional parametr specifying path to the eye data. If missing, default value from global variables is used.}
}
\description{
Saves an eye object as RData file into directory with eye data
}
\author{
Filip Dechterenko
}

