% Generated by roxygen2 (4.0.1): do not edit by hand
\name{scanpath.to.space}
\alias{scanpath.to.space}
\title{Converts scanpath to space}
\usage{
scanpath.to.space(spath, min.x = NULL, min.y = NULL, min.t = 0)
}
\arguments{
\item{spath}{scanpath, which will be converted}

\item{min.x}{minimum value in x coordinate, default value is set from config}

\item{min.y}{minimum value in y coordinate, default value is set from config}

\item{min.t}{minimum value inty coordinate, default value is 0}
}
\description{
Converts scanpath to disrete spatiotemporal map
}
\author{
Filip Dechterenko
}

