% Generated by roxygen2 (4.0.1): do not edit by hand
\name{smooth.space}
\alias{smooth.space}
\title{Convolves scanpath with Gaussian}
\usage{
smooth.space(spath, G, checks = T)
}
\arguments{
\item{spath}{scanpath, which will be convolved}

\item{G}{small Gaussian used for convlution}

\item{checks}{boolean, if we should run tests. We might want to set this to F to reduce running time. Default value is T}
}
\description{
Convolves scanpath with Gaussian. First, we bin scanpath to create discrete spatiotemporal map (stm).
We add Gaussians to each position at stm
}
\author{
Filip Dechterenko
}

