% Generated by roxygen2 (4.0.1): do not edit by hand
\name{step.size}
\alias{step.size}
\title{Computes step size per dimension}
\usage{
step.size(steps)
}
\arguments{
\item{steps}{one dimension of object steps}
}
\description{
Computes step size per one dimension of object steps
}
\author{
Filip Dechterenko
}

