% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sum.spaces}
\alias{sum.spaces}
\title{Sum spatiotemporal fixation maps}
\usage{
\method{sum}{spaces}(SVs)
}
\arguments{
\item{SVs}{list of spatiotemporal fixation maps}
}
\description{
Sum spatiotemporal fixation maps
}
\author{
Filip Dechterenko
}

