% Generated by roxygen2 (4.0.1): do not edit by hand
\name{valueToIndex}
\alias{valueToIndex}
\title{Gets index for given value}
\usage{
valueToIndex(x, min.x, step.x)
}
\arguments{
\item{x}{value to be converted to index}

\item{min.x}{minimum value of x}

\item{step.x}{stepsize of x}
}
\description{
Gets index for given value based on minimum value and stepsize
}
\author{
Filip Dechterenko
}

